requirements.txt
gradio>=5.0
anthropic
openai
python-dotenv
Pillow
